Option to fail in case of duplicate operationIds
Added an option exposed to configuraion of the SwaggerModule. This can be
useful eg. in the JAX-RS scenario with code-first generation of docs,
where it is possible to inadvertently add a method with the same name
in different resources. This by default generates a duplicate operationId,
but today the OperationIdForMethod is fixing it automatically by making
the name unique with a suffix.

In certain cases it is useful to fail, eg. a Maven build, and
let the API authors take proper action to disambiguate operationIds.